+++ src/Opaleye/Column.hs
@@ -16,13 +16,12 @@ module Opaleye.Column (-- * 'Column'
                        maybeToNullable,
                        -- * Unsafe operations
                        unsafeCast,
-                       unsafeCoerce,
                        unsafeCoerceColumn,
                        unsafeCompositeField,
                        -- * Entire module
                        module Opaleye.Column)  where
 
-import           Opaleye.Internal.Column (Column, Nullable, unsafeCoerce, unsafeCoerceColumn,
+import           Opaleye.Internal.Column (Column, Nullable, unsafeCoerceColumn,
                                           unsafeCast, unsafeCompositeField)
 import qualified Opaleye.Internal.Column as C
 import qualified Opaleye.Internal.HaskellDB.PrimQuery as HPQ
@@ -50,7 +49,8 @@ matchNullable replacement f x = C.unsafeIfThenElse (isNull x) replacement
 -- | If the @Column (Nullable a)@ is NULL then return the provided
 -- @Column a@ otherwise return the underlying @Column a@.
 --
--- The Opaleye equivalent of 'Data.Maybe.fromMaybe'.
+-- The Opaleye equivalent of 'Data.Maybe.fromMaybe' and very similar
+-- to PostgreSQL's @COALESCE@.
 fromNullable :: Column a -> Column (Nullable a) -> Column a
 fromNullable = flip matchNullable id