+++ src/Opaleye/Internal/PackMap.hs
@@ -40,7 +40,7 @@ import qualified Data.Functor.Identity as I
 -- 'ProductProfunctor') in @s@ and @t@.  It is unclear at this point
 -- whether we want the same @Traversal@ laws to hold or not.  Our use
 -- cases may be much more general.
-data PackMap a b s t = PackMap (forall f. Applicative f => (a -> f b) -> s -> f t)
+newtype PackMap a b s t = PackMap (forall f. Applicative f => (a -> f b) -> s -> f t)
 
 -- | Replaces the targeted occurences of @a@ in @s@ with @b@ (changing
 -- the @s@ to a @t@ in the process).  This can be done via an
@@ -114,6 +114,11 @@ eitherFunction :: Functor f
                -> f (Either b b')
 eitherFunction f g = fmap (either (fmap Left) (fmap Right)) (f PP.+++! g)
 
+-- | Like 'Control.Lens.Iso.iso'.  In practice it won't actually be
+-- used as an isomorphism, but it seems to be appropriate anyway.
+iso :: (s -> a) -> (b -> t) -> PackMap a b s t
+iso h g = PackMap (dimap h (fmap g))
+
 -- {
 
 -- Boilerplate instance definitions.  There's no choice here apart
@@ -134,13 +139,6 @@ instance ProductProfunctor (PackMap a b) where
   (***!) = PP.defaultProfunctorProduct
 
 instance PP.SumProfunctor (PackMap a b) where
-  f +++! g =
-    PackMap (\x ->
-               case f of
-                 PackMap f' ->
-                   case g of
-                     PackMap g' ->
-                       eitherFunction (f' x)
-                                      (g' x))
+  PackMap f +++! PackMap g = PackMap (\x -> eitherFunction (f x) (g x))
 
 -- }