+++ src/Opaleye/Internal/TableMaker.hs
@@ -5,6 +5,7 @@ module Opaleye.Internal.TableMaker where
 import qualified Opaleye.Column as C
 import qualified Opaleye.Internal.Column as IC
 import qualified Opaleye.Internal.PackMap as PM
+import qualified Opaleye.Internal.Unpackspec as U
 
 import           Data.Profunctor (Profunctor, dimap)
 import           Data.Profunctor.Product (ProductProfunctor, empty, (***!))
@@ -21,18 +22,19 @@ import qualified Opaleye.Internal.HaskellDB.PrimQuery as HPQ
 newtype ViewColumnMaker strings columns =
   ViewColumnMaker (PM.PackMap () () strings columns)
 
-newtype ColumnMaker columns columns' =
-  ColumnMaker (PM.PackMap HPQ.PrimExpr HPQ.PrimExpr columns columns')
-
 runViewColumnMaker :: ViewColumnMaker strings tablecolumns ->
                        strings -> tablecolumns
 runViewColumnMaker (ViewColumnMaker f) = PM.overPM f id
 
+{-# DEPRECATED ColumnMaker "Use Unpackspec instead" #-}
+type ColumnMaker = U.Unpackspec
+
+{-# DEPRECATED runColumnMaker "Use runUnpackspec instead" #-}
 runColumnMaker :: Applicative f
                   => ColumnMaker tablecolumns columns
                   -> (HPQ.PrimExpr -> f HPQ.PrimExpr)
                   -> tablecolumns -> f columns
-runColumnMaker (ColumnMaker f) = PM.traversePM f
+runColumnMaker = U.runUnpackspec
 
 -- There's surely a way of simplifying this implementation
 tableColumn :: ViewColumnMaker String (C.Column a)
@@ -40,16 +42,12 @@ tableColumn = ViewColumnMaker
               (PM.PackMap (\f s -> fmap (const (mkColumn s)) (f ())))
   where mkColumn = IC.Column . HPQ.BaseTableAttrExpr
 
-column :: ColumnMaker (C.Column a) (C.Column a)
-column = ColumnMaker
-         (PM.PackMap (\f (IC.Column s)
-                      -> fmap IC.Column (f s)))
-
 instance Default ViewColumnMaker String (C.Column a) where
   def = tableColumn
 
-instance Default ColumnMaker (C.Column a) (C.Column a) where
-  def = column
+{-# DEPRECATED column "Use unpackspecColumn instead" #-}
+column :: ColumnMaker (C.Column a) (C.Column a)
+column = U.unpackspecColumn
 
 -- {
 
@@ -66,20 +64,6 @@ instance Profunctor ViewColumnMaker where
   dimap f g (ViewColumnMaker q) = ViewColumnMaker (dimap f g q)
 
 instance ProductProfunctor ViewColumnMaker where
-  empty = PP.defaultEmpty
-  (***!) = PP.defaultProfunctorProduct
-
-instance Functor (ColumnMaker a) where
-  fmap f (ColumnMaker g) = ColumnMaker (fmap f g)
-
-instance Applicative (ColumnMaker a) where
-  pure = ColumnMaker . pure
-  ColumnMaker f <*> ColumnMaker x = ColumnMaker (f <*> x)
-
-instance Profunctor ColumnMaker where
-  dimap f g (ColumnMaker q) = ColumnMaker (dimap f g q)
-
-instance ProductProfunctor ColumnMaker where
   empty = PP.defaultEmpty
   (***!) = PP.defaultProfunctorProduct